# REST APIs Implementation/api/openapi.yaml
openapi: 3.0.1
info:
  title: Film Manager
  description: REST APIs for the interaction with the Film Manager service.
  termsOfService: http://swagger.io/terms/
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
servers:
  - url: https://todo.swagger.io/v2
  - url: http://todo.swagger.io/v2
...
paths:
  /api/films:
    post:
      tags:
        - films
      summary: Create a new film
      description: A new film is created by the authenticated user.
      operationId: createFilm
      requestBody:
        description: Representation of the film to be created
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Film"
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Film"
        "400":
          description: Bad Request
          content: {}
        "401":
          description: Unauthorized
          content: {}
        "500":
          description: Internal Server Error
          content: {}
      security:
        - cookieAuth: []
      x-swagger-router-controller: Films
  /api/films/private:
    get:
      tags:
        - films/private
      summary: Retrieve the private films of the logged-in user
      description:
        The private films of the logged-in user are retrieved. A pagination
        mechanism is implemented to limit the size of messages.
      operationId: getPrivateFilms
      parameters:
        - name: pageNo
          in: query
          description: "The id of the requested page (if absent, the first page is returned)"
          required: false
          style: form
          explode: true
          schema:
            type: integer
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/inline_response_200"
        "401":
          description: Unauthorized
          content: {}
        "500":
          description: Internal Server Error
          content: {}
      security:
        - cookieAuth: []
      x-swagger-router-controller: Filmsprivate
  /api/films/private/{filmId}:
    get:
      tags:
        - "films/private/{filmId}"
      summary: Retrieve a private film
      description:
        The private film with ID filmId is retrieved. This operation can
        be performed on the film if the user who performs the operation is the film's
        owner.
      operationId: getSinglePrivateFilm
      parameters:
        - name: filmId
          in: path
          description: ID of the film to retrieve
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Film"
        "401":
          description: Unauthorized
          content: {}
        "403":
          description: Forbidden
          content: {}
        "404":
          description: Not Found
          content: {}
        "500":
          description: Internal Server Error
          content: {}
      security:
        - cookieAuth: []
      x-swagger-router-controller: FilmsprivatefilmId
    put:
      tags:
        - "films/private/{filmId}"
      summary: Update a private film
      description:
        The private film with ID filmId is updated. This operation does
        not allow changing its visibility. This operation can be performed only by
        the owner.
      operationId: updateSinglePrivateFilm
      parameters:
        - name: filmId
          in: path
          description: ID of the film to update
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      requestBody:
        description: The updated film object that needs to replace the old object
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Film"
        required: true
      responses:
        "204":
          description: No Content
          content: {}
        "400":
          description: Bad Request
          content: {}
        "401":
          description: Unauthorized
          content: {}
        "403":
          description: Forbidden
          content: {}
        "404":
          description: Not Found
          content: {}
        "409":
          description: "Conflict (attempt to make unallowed change, e.g., visibility)"
          content: {}
        "500":
          description: Internal Server Error
          content: {}
      security:
        - cookieAuth: []
      x-swagger-router-controller: FilmsprivatefilmId
    delete:
      tags:
        - "films/private/{filmId}"
      summary: Delete a private film
      description:
        The private film with ID filmId is deleted. This operation can
        only be performed by the owner.
      operationId: deleteSinglePrivateFilm
      parameters:
        - name: filmId
          in: path
          description: ID of the film to delete
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: Successful operation - No Content
          content: {}
        "401":
          description: Unauthorized
          content: {}
        "403":
          description: Forbidden
          content: {}
        "404":
          description: Not Found
          content: {}
        "500":
          description: Internal Server Error
          content: {}
      security:
        - cookieAuth: []
      x-swagger-router-controller: FilmsprivatefilmId
  /api/films/public:
    get:
      tags:
        - films/public
      summary: Retrieve the public films
      description:
        "The public films (i.e., the films that are visible for all the\
        \ users of the service) are retrieved. This operation does not require authentication.\
        \ A pagination mechanism is implemented to limit the size of messages."
      operationId: getPublicFilms
      parameters:
        - name: pageNo
          in: query
          description: "The id of the requested page (if absent, the first page is returned)"
          required: false
          style: form
          explode: true
          schema:
            type: integer
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/inline_response_200"
        "500":
          description: Internal Server Error
          content: {}
      x-swagger-router-controller: Filmspublic
  /api/films/public/invited:
    get:
      tags:
        - films/public/invited
      summary:
        "Retrieve the public films that the logged-in user has been invited\
        \ to review, possibly filtered by the review's invitation Status"
      description:
        "The public films that the logged-in user has been invited to review\
        \ are retrieved, possibly filtered by the review's invitation Status. A pagination\
        \ mechanism is implemented to limit the size of messages."
      operationId: getInvitedFilms
      parameters:
        - name: invitationStatus
          in: query
          description:
            "The invitation Status used to filter the reviews (if absent,\
            \ all reviews are considered)"
          required: false
          style: form
          explode: true
          schema:
            type: string
            enum:
              - PENDING
              - ACCEPTED
              - REFUSED
        - name: pageNo
          in: query
          description: "The id of the requested page (if absent, the first page is returned)"
          required: false
          style: form
          explode: true
          schema:
            type: integer
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/inline_response_200"
        "401":
          description: Unauthorized
          content: {}
        "404":
          description: Not found
          content: {}
        "500":
          description: Internal Server Error
          content: {}
      security:
        - cookieAuth: []
      x-swagger-router-controller: Filmspublicinvited
  /api/films/public/{filmId}:
    get:
      tags:
        - "films/public/{filmId}"
      summary: Retrieve a public film
      description:
        The public film with ID filmId is retrieved. This operation does
        not require authentication.
      operationId: getSinglePublicFilm
      parameters:
        - name: filmId
          in: path
          description: ID of the film to retrieve
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Film"
        "404":
          description: Not Found
          content: {}
        "500":
          description: Internal Server Error
          content: {}
      x-swagger-router-controller: FilmspublicfilmId
    put:
      tags:
        - "films/public/{filmId}"
      summary: Update a public film
      description:
        The public film with ID filmId is updated. This operation does
        not allow changing its visibility. This operation can be performed only by
        the owner.
      operationId: updateSinglePublicFilm
      parameters:
        - name: filmId
          in: path
          description: ID of the film to update
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      requestBody:
        description: The updated film object that needs to replace the old object
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Film"
        required: true
      responses:
        "204":
          description: No Content
          content: {}
        "400":
          description: Bad Request
          content: {}
        "401":
          description: Unauthorized
          content: {}
        "403":
          description: Forbidden
          content: {}
        "404":
          description: Not Found
          content: {}
        "409":
          description: "Conflict (attempt to make unallowed change, e.g., visibility)"
          content: {}
        "500":
          description: Internal Server Error
          content: {}
      security:
        - cookieAuth: []
      x-swagger-router-controller: FilmspublicfilmId
    delete:
      tags:
        - "films/public/{filmId}"
      summary: Delete a public film
      description:
        The public film with ID filmId is deleted. This operation can only
        be performed by the owner.
      operationId: deleteSinglePublicFilm
      parameters:
        - name: filmId
          in: path
          description: ID of the film to delete
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: Successful operation - No Content
          content: {}
        "401":
          description: Unauthorized
          content: {}
        "403":
          description: Forbidden
          content: {}
        "404":
          description: Not Found
          content: {}
        "500":
          description: Internal Server Error
          content: {}
      security:
        - cookieAuth: []
      x-swagger-router-controller: FilmspublicfilmId
...
components:
  schemas:
...
    Film:
      required:
        - owner
        - private
        - title
      type: object
      properties:
        $schema:
          type: string
          description:
            path to the JSON Schema to use for the validation of the JSON
            data structure describing the film
        id:
          type: integer
          description: unique identifier for the film data structure
        title:
          type: string
          description: textual title of the film
        owner:
          type: integer
          description: unique identifier of the the film data structure owner
        private:
          type: boolean
          description:
            "true if only the owner of the film data structure can access\
            \ it, false otherwise"
          default: true
        watchDate:
          type: string
          description: the date when the film was watched by the owner
          format: date
        rating:
          maximum: 10
          minimum: 1
          type: integer
          description: the rating the owner has given to the film
        favorite:
          type: boolean
          description:
            "true if the film is among the favourite ones of the owner,\
            \ false otherwise "
          default: false
        self:
          type: string
          format: uri
      additionalProperties: false
      description: schema for Film data structures
      example:
        owner: 6
        private: true
        watchDate: 2000-01-23
        $schema: $schema
        rating: 2
        self: http://example.com/aeiou
        id: 0
        title: title
        favorite: false
...
    inline_response_200:
      type: object
      properties:
        totalPages:
          type: integer
        currentPage:
          type: integer
        totalItems:
          type: integer
        films:
          type: array
          items:
            $ref: "#/components/schemas/Film"
        next:
          type: string
          format: uri
      example:
        next: http://example.com/aeiou
        films:
          - owner: 6
            private: true
            watchDate: 2000-01-23
            $schema: $schema
            rating: 2
            self: http://example.com/aeiou
            id: 0
            title: title
            favorite: false
          - owner: 6
            private: true
            watchDate: 2000-01-23
            $schema: $schema
            rating: 2
            self: http://example.com/aeiou
            id: 0
            title: title
            favorite: false
        totalItems: 1
        totalPages: 0
        currentPage: 6
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      description: JWT to identify the authenticated user.
      name: connect.sid
      in: cookie